requirements.txt
flask>=3.0
cryptography>=42.0
pystray>=0.19
Pillow>=10.0
pyperclip>=1.8
